requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
PyYAML>=6.0.1
matplotlib>=3.4.3
pyvisgraph>=0.2.1
tqdm>=4.65.0
docutils>=0.20.1
ipywidgets>=8.0.0